Release v0.1.43

+++ concordium-node/src/common/serialization/serializable.rs
@@ -177,6 +177,8 @@ where
 // Concordium-common
 // ==============================================================================================
 
+/// As `UCursor` maintains compatability with a broad range of `std::io`, we
+/// simply enforce a max size of u32 only during serialization-phases.
 impl Serializable for UCursor {
     /// It makes a `deep-copy` of the `UCursor` into `Archive`.
     fn serialize<A>(&self, archive: &mut A) -> Fallible<()>
@@ -185,7 +187,7 @@ impl Serializable for UCursor {
         let mut self_from = self.sub(self.position())?;
         let self_from_len = self_from.len();
 
-        self_from_len.serialize(archive)?;
+        (self_from_len as u32).serialize(archive)?;
         std::io::copy(&mut self_from, archive)?;
 
         Ok(())
